requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
oauthlib==3.2.2
pyOpenSSL==23.3.0

# Task Queue and Caching
celery==5.3.6
redis==5.0.1
flower==2.0.1

# API Documentation
flasgger==0.9.7.1
apispec==6.3.1

# Utils
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
validators==0.22.0
phonenumbers==8.13.26
pycountry==23.12.11
babel==2.14.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==21.0.0

# Production
gunicorn==21.2.0
supervisor==4.2.5 